+++ web_src/less/index.less
@@ -13,6 +13,7 @@
 @import "./markup/content.less";
 @import "./markup/codecopy.less";
 @import "./code/linebutton.less";
+@import "./markup/asciicast";
 
 @import "./chroma/base.less";
 @import "./chroma/light.less";
@@ -36,6 +37,5 @@
 @import "_explore";
 @import "_review";
 @import "_package";
-@import "_asciicast";
 
 @import "./helpers.less";
