+++ web_src/less/markup/asciicast.less
@@ -1,4 +1,4 @@
-@import "asciinema-player/dist/bundle/asciinema-player.css";
+@import "../asciinema-player/dist/bundle/asciinema-player.css";
 
 .asciinema-player-container {
   width: 100%;
